Add tests that use the FlameGraph test files
The main functionality of collapse-perf has been split out into a library to
allow the tests to call handle_file() directly and get a result without
creating temporary files.

Flamegraph has been added as a git submodule so we can use their test files in
flamegraph/test. Not all tests are passing so those still need to be looked
into.