Add tests that use the FlameGraph test files (#11)
Add tests that use the FlameGraph test files

The main functionality of `collapse-perf` has been split out into a
library to allow the tests to call `handle_file()` directly and get a
result without creating temporary files.

FlameGraph has been added as a git submodule so we can use their test
files in `flamegraph/test`. Not all tests are passing so those still
need to be looked into. One case that was caught was that the perl
version handles stack lines without a function name, like

    7f1e2215d058  (/lib/x86_64-linux-gnu/libc-2.15.so)

by setting the function name to " "; this PR matches that behavior,
which makes 5 more tests pass.