Add un-inline support using addr2line (#14)
This patch adds full support for the `--inline` and `--context` flags of
`stackcollapse-perf` by relying on the `addr2line` crate to resolve symbol
addresses instead of the names reported by perf, which _can_ lead to more
informative frame names.

Note that, on modern systems, address-space randomization means that we often
do not have the "true" address of functions in the trace from perf script. This
issue also plagues stackcollapse-perf. There are some possible workarounds
(#31), but nothing definite as of yet. For the time being we're aiming for
parity.

This patch also adds another set of tests that are _not_ from the upstream
FrameGraph repository. Specifically, it compares the output of
`stackcollapse-perf` with `--inline` and `--context` with ours on a simple
binary that we've compiled and profiled without randomization.

This patch _also_ adopts the `log` crate, and emits appropriate log messages
whenever the input is malformed in some way. We explicitly do not return with
an error, since we want to be liberal in what we accept (just like
`stackcollapse-perf`).