Replace pretty_toa with num-format.
Unlike pretty_toa, num-format can create
a &str representation of a number (with
thousands separators) without allocating.

Creating this kind of string representation from
a number only occurs at one place in inferno's
code; so replacing pretty_toa with num-format
is a tiny improvement, but it is an improvement
nontheless as it avoids an unnecessary allocation
(which occurs in a loop).

Fixes #29.