Introduce "Collapser" abstraction to simplify code for binaries.
Since hopefully inferno will eventually have as many binary
applications as the original perl version (one for each "frontend"
such as perf and bpftrace), this is an initial attempt at creating
an abstraction that will make writing all those very similar
applications easier.

Also - Unrelated to the above, this commit incorporates a suggested
change to the test in bptrace.rs.