Add support for differential flame graphs (#60)
This adds the ability to easily compare before and after profiles when
benchmarking code changes. If the folded profile file has an extra
samples column at the end it will be interpreted as a differential
profile and the delta between the before and after will be calculated.
The resulting flame graph is drawn using the after profile, but is
colorized based on the delta. It gets a red hue where the after profile
spends more time, and a blue hue where it spends less time. You can
also pass the --negate flag to switch the red and blue.

To create the differential profile files that this feature expects, you
can use the difffolded.pl script from upstream FlameGraph. We have not
yet ported this to Rust.