Implement --factor option (#65)
This option allows you to pass in a factor to scale sample counts by in
the flame graph.

It can be useful if the sample data has fractional sample counts since
the fractional parts are stripped off when creating the flame graph. To
work around this you can scale up the sample counts to be integers, then
scale them back down in the graph by passing in a scale factor.