Fix coverage timeouts caused by cli tests (#85)
This PR removes the tests that call out to the CLIs and instead test much of
what the main functions did by exposing those functions as part of the library.

I've added `from_files` to `flamegraph` and moved most of the code from the
main functions over to it. The tests now use this instead of calling
`from_reader` or `from_readers` directly.

The collapse tests now also call `collapse_file` instead of calling `collapse`
directly, unless the file is a gzip. I had to change `collapse_file` to take a
`Write` instead of always writing to `STDOUT` so it could be used in tests.