Fix bug with &str defaults in CLI options (#104)
For all of the default constants, two different versions are created.
One in `flamegraph::defaults` and the other in
`flamegraph::defaults::str`. Unfortunately, because of the way the macro
for defining the constants works, if the type of the constant is `&str`, the
version in `flamegraph::defaults::str` includes the quotes of the string
literal in the string constant.

To fix this, I used the version of the constants in `flamegraph::defaults` for
all of the &str constants.