Ensure reliable nameattr ordering
Hashbrown has a different ordering from HashMap, even with deterministic
fnv hashing, so the tests were failing on beta. This changes nameattr to
use an indexmap instead, which has a reliable ordering at all times
(insertion order). This is probably a better behavior anyway, because it
means that attributes will be added in the order they appear in the
nameattr file.