Restructure multithreading -> 1 pass over data.
This commit restructures multithreading so that only one pass is made
over the data (instead of the previous two).

Also, since I had to add some command line options (`nstacks` and
`nthreads`), I figured I would also do an additional cleanup on how
we're doing the cli. The order in which options show up in `--help`
is set according to an alphabetical ordering of the field names in
the `Opt` struct. The actual values the user sees, however, are given
by the `value_name` attribute on the struct field. I did a little
renaming of the struct fields (which doesn't affect what the user
sees) in order to have the options show up in alphabetical order
when the user runs `--help`.