Make harder to forget stuff when change defaults
The `nthreads` and `nstacks` default values are now in just one place in
the code, which will make it much harder to forget about places they
need to be updated if we ever want to change them. One would still have
to update the doc comments if you changed the default values, but I
don't see anyway around that. This at least makes the situation a little
better, i.e. the code more resilient.