Generalize tests and fix bug in nstacks bench.
This commit generalizes the tests, moving a bunch of functionality out
of `collapse::perf` and into `collapse` so that it can be used for all
collapsers.

It also fixes a bug in the code that was benchmarking the
`nstacks_per_job` paramenter. After fixing this bug, it looks like, on
my machine at least, 100 is a good default for `nstacks_per_job`.

Finally, I wondered if we could get a bit of a speed up if we used
`Vec::with_capacity` instead of `Vec::new` for the buffer that being
filled up and sent across threads. It seems we get a bit of a speed
boost if we pre-allocate approximately 1024 * nstack_per_job number of
bytes.