Same thing for perf ...
... collapse_single_threaded should return an error if it's state is not
set back to the initial state once it returns because this means we
stopped processing the middle of a stack. Now that
collapse_single_threaded returns an error in the case that it's state
has not been reset, we can rely on the state being clean after each call
to it and no longer have to clear the state in other places.