Actuall, perf always "ends in the middle...
... of a stack" because what tells us that we need to start a new stack is
an empty line and apparently perf files don't need to end in an
empty line. So we can't say "At the end of a file, perf should always
have a clean state otherwise it's an error."

So I reverted the change that had perf's collapse_single_threaded return
an error if it had state when that method exits. Instead, I made it so
whenever collapse_single_threaded exits, the state is reset by us.