Stop work if error / scrutinize for deadlocks
Here is a change that restructures our multithreaded code so that we can
stop doing unecessary work if we encounter an error. Before, we would
continue processing all the data even if the first stack produced an
error.

I only made this change for dtrace because I think I'm going to tackle
creating an abstaction so all of our Folders can share more code
next.