Correctly annotate jitdump stack frames
When using perf with the modern jitdump format, `perf inject` creates an
ELF object for each JITted function and replaces the function entries in
the recording to point to those SOs. This causes `stackcollapse-perf` to
not recognise the functions as the module is no longer displayed as
being the perf JIT map for the process.