Add a generic palette to highlight kernel / jitted code
The "java" and "js" colour palettes have a very useful feature of
being able to highlight which frames are jitted or kernel functions
based on the annotations added to the function makes by the
stackcollapse programs. Unfortunately they also make a number of
assumptions based on the function names which are very specific to
those languages, and cause nonsensical results with other runtimes.

This PR adds a new generic "annotated" colour palette that only uses
the function name annotations, and can thus be used regardless of
the runtime being profiled.