Increase default flamegraph fidelity
See also #153. PR #136 changed the semantics of the `--minwidth` option
from output pixel width to percentage width (which is effectively the
same as pct of total samples), which greatly reduced the default
fidelity of flamegraphs as 0.1% of total time is quite coarse,
especially once you start zooming in.

This PR reduces the default to 0.01%, which produces pretty much
identical output as flamegraph.pl / old inferno at ~1200px wide and
includes a much greater number of bars. I think this change is worth
making as it makes it reduces friction for people migrating from
flamegraph.pl, and in my tests produces much more useful output by
default.

I've also updated the description of the option to match the current
behaviour, as it still mentioned pixels and that had me confused for
a good while.