Upgrade to quick-xml 0.27
Note that since we have `pub` functions that return `quick_xml` types,
this would be a breaking change.

Interestingly enough, we only expose `quick_xml` in the form of
`quick_xml::Result` as far as I can tell. Specifically, we're using its
error type, which feels like something we can avoid to mitigate this
kind of "transitive bump" in the future. In particular, I think we could
reasonable start using an opaque error type for these methods instead.